+++ amuyu/lib/screens/home_screen.dart
@@ -5,6 +5,7 @@ import 'package:amuyu/screens/family_tree_screen.dart';
 import 'package:amuyu/screens/historical_log_screen.dart';
 import 'package:amuyu/widgets/menu_card.dart';
 import 'package:flutter/material.dart';
+import 'package:amuyu/screens/family_documents_screen.dart';
 
 class HomeScreen extends StatelessWidget {
   const HomeScreen({super.key});
@@ -46,6 +47,16 @@ class HomeScreen extends StatelessWidget {
               );
             },
           ),
+
+ MenuCard(
+            title: 'Documentos Familiares',
+            icon: Icons.folder_copy, // Un buen ícono para documentos
+            onTap: () {
+              Navigator.of(context).push(
+                MaterialPageRoute(builder: (_) => const FamilyDocumentsScreen()),
+              );
+    },
+    ),
         ],
       ),
     );